DJANGO/PeluqueriaApp/requirements.txt
django == 4.1
django-admin-interface
django-ckeditor
django-crispy-forms
django-import-export
django-paranoid
django-rest-framework
pillow
